+++ custom_components/prix_carburant/sensor.py
@@ -4,13 +4,19 @@ from __future__ import annotations
 from datetime import datetime, timedelta
 import logging
 
+import voluptuous as vol
+
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity
-from homeassistant.config_entries import ConfigEntry
+from homeassistant.config_entries import SOURCE_IMPORT, ConfigEntry
 from homeassistant.const import ATTR_LATITUDE, ATTR_LONGITUDE, ATTR_NAME, CURRENCY_EURO
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
+from homeassistant.helpers.aiohttp_client import async_get_clientsession
+import homeassistant.helpers.config_validation as cv
+from homeassistant.helpers.config_validation import PLATFORM_SCHEMA_BASE
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .const import (
@@ -24,12 +30,36 @@ from .const import (
     ATTR_UPDATED_DATE,
     CONF_FUELS,
     CONF_MAX_KM,
+    CONF_STATIONS,
     DOMAIN,
     FUELS,
 )
 from .tools import PrixCarburantTool
 
 _LOGGER = logging.getLogger(__name__)
+
+# Validation of the yaml configuration
+PLATFORM_SCHEMA = PLATFORM_SCHEMA_BASE.extend(
+    {
+        vol.Optional(CONF_STATIONS, default=[]): cv.ensure_list,
+    }
+)
+
+
+async def async_setup_platform(
+    hass: HomeAssistant,
+    config: ConfigType,
+    async_add_entities: AddEntitiesCallback,
+    discovery_info: DiscoveryInfoType | None = None,
+) -> None:
+    """Set up the Prix Carburant sensor."""
+    hass.async_create_task(
+        hass.config_entries.flow.async_init(
+            DOMAIN,
+            context={"source": SOURCE_IMPORT},
+            data=config,
+        )
+    )
 
 
 async def async_setup_entry(
@@ -40,18 +70,30 @@ async def async_setup_entry(
     """Set up the platform from config_entry."""
     config = entry.data
     options = entry.options
-    max_distance = options.get(CONF_MAX_KM, config.get(CONF_MAX_KM))
+
+    websession = async_get_clientsession(hass)
+
+    tool = PrixCarburantTool(time_zone=hass.config.time_zone, session=websession)
+
+    # yaml configuration
+    if CONF_STATIONS in config:
+        _LOGGER.info("Init stations data from yaml list")
+        await tool.init_stations_from_list(config[CONF_STATIONS])
+    # ui configuration
+    else:
+        _LOGGER.info("Init stations list near Home-Assistant location")
+        max_distance = options.get(CONF_MAX_KM, config.get(CONF_MAX_KM))
+        await tool.init_stations_from_location(
+            user_latitude=hass.config.latitude,
+            user_longitude=hass.config.longitude,
+            user_range=max_distance,
+        )
+        _LOGGER.info("%s stations found", str(len(tool.stations)))
+
     enabled_fuels = {}
     for fuel in FUELS:
         fuel_key = f"{CONF_FUELS}_{fuel}"
         enabled_fuels[fuel] = options.get(fuel_key, config.get(fuel_key, True))
-
-    tool = PrixCarburantTool(
-        hass.config.latitude, hass.config.longitude, max_distance, hass.config.time_zone
-    )
-    _LOGGER.info("Get stations list near Home-Assistant location")
-    await tool.init_stations_data()
-    _LOGGER.info("%s stations found", str(len(tool.stations)))
 
     async def async_update_data():
         """Fetch data from API."""
@@ -63,7 +105,7 @@ async def async_setup_entry(
         _LOGGER,
         name=DOMAIN,
         update_method=async_update_data,
-        update_interval=timedelta(seconds=120),
+        update_interval=timedelta(minutes=60),
     )
 
     await coordinator.async_refresh()
@@ -103,6 +145,19 @@ class PrixCarburant(SensorEntity):
         else:
             station_name = f"Station {self.station_id}"
         self._attr_name = f"{station_name} {self.fuel}"
+
+        match self.station_info[ATTR_BRAND]:
+            case "Système U":
+                self._attr_entity_picture = "https://upload.wikimedia.org/wikipedia/fr/1/13/U_commer%C3%A7ants_logo_2018.svg"
+            case "Total":
+                self._attr_entity_picture = "https://upload.wikimedia.org/wikipedia/fr/f/f7/Logo_TotalEnergies.svg"
+            case "Total Access":
+                self._attr_entity_picture = "https://upload.wikimedia.org/wikipedia/fr/f/f7/Logo_TotalEnergies.svg"
+            case "Intermarché":
+                self._attr_entity_picture = "https://upload.wikimedia.org/wikipedia/fr/8/8c/Logo_Groupe_Les_Mousquetaires.svg"
+            case "Leclerc":
+                self._attr_entity_picture = "https://upload.wikimedia.org/wikipedia/commons/e/ed/Logo_E.Leclerc_Sans_le_texte.svg"
+
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, self.station_id)},
             manufacturer="Station",